support audio recitation
- internal recitations (the same as Recite).
- external recitations (a custom url), but only from the url params.
- auto-reciting before an aaya or after an aaya.
- reciting on-demand (by clicking on aaya (but not on its number).
  (on-demand is enabled with auto-reciting, but not in Insert mode.)
- preload the next aaya in auto-reciting mode.
- ctrl+space to repeat the current aaya recitation.
- audio url params: qari= (id or custom url), qs= ("a" or "b"), qp=.
- other url params: nohelp, help (default); nostats, stats (default).
- make: support toggle minifications.
- split global state from script.js

+++ make.sh
@@ -4,15 +4,18 @@ set -eu
 minify_js=true
 minify_css=true
 minify_html=true
-
-# minify_js=false
-# minify_css=false
-# minify_html=false
-
 disabledark=true
 # remove the darkmode js & css, because it's still experimental and thus is disabled by default
 
-# disabledark=false
+# use `make toggle`
+if [ -e .nominify ]; then
+  minify_js=false
+  minify_css=false
+  minify_html=false
+fi
+if [ -e .keepdark ]; then
+  disabledark=false
+fi
 
 _() { printf '%s\n' "$*"; "$@"; }
 
@@ -20,7 +23,9 @@ dist=false
 if [ $# -gt 0 ]; then
   case "$1" in clean|-B|-f)
     _ rm -f index.html .minified.{script,style}
-    _ rm -f .hashes_vars.sh mymeta.json.zst
+    if [ "$1" != -f ]; then
+      _ rm -f .hashes_vars.sh mymeta.json.zst
+    fi
     if [ -e irsaa.html ]; then
       _ rm -f irsaa.html .minified2.{script,style} LOAD.js changelog.html
     fi
@@ -45,7 +50,7 @@ source .hashes_vars.sh
 
 # JAVASCRIPT MINIFICATION
 
-js_opts=(--compress passes=10,toplevel --mangle toplevel --mangle-props "reserved=[$reserved_props]")
+js_opts=(--compress passes=10,toplevel,drop_console --mangle toplevel --mangle-props "reserved=[$reserved_props]")
 # js_opts=(--compress)
 
 js_process() { perl -pe '
@@ -147,7 +152,7 @@ needed() {
   return 1
 }
 
-jss='globalsfree.js init.js load.js aftermeta.js osk.js script.js main.js'
+jss='globalsfree.js init.js load.js aftermeta.js osk.js state.js audio.js script.js main.js'
 
 if needed .minified.script -- $jss .hashes_vars.sh; then
   printf 'Preparing %s... ' 'the script'
